Fix #5135: Move well-known names to a new object `WellKnownNames`.
This breaks the circular dependency between the initialization of
`object Types` and `object Names`.
