+++ ir/shared/src/test/scala/org/scalajs/ir/TestIRBuilder.scala
@@ -19,6 +19,7 @@ import OriginalName.NoOriginalName
 import Printers._
 import Trees._
 import Types._
+import WellKnownNames._
 
 object TestIRBuilder {
