+++ linker/jvm/src/test/scala/org/scalajs/linker/RunTest.scala
@@ -24,6 +24,7 @@ import org.junit.rules.TemporaryFolder
 import org.scalajs.ir.Names._
 import org.scalajs.ir.Trees._
 import org.scalajs.ir.Types._
+import org.scalajs.ir.WellKnownNames._
 
 import org.scalajs.junit.async._
