+++ linker/shared/src/main/scala/org/scalajs/linker/backend/emitter/NameGen.scala
@@ -18,6 +18,7 @@ import org.scalajs.ir.{OriginalName, UTF8String}
 import org.scalajs.ir.Names._
 import org.scalajs.ir.OriginalName.NoOriginalName
 import org.scalajs.ir.Types._
+import org.scalajs.ir.WellKnownNames._
 
 /** Performs state independent name mangling.
  *
