+++ linker/shared/src/main/scala/org/scalajs/linker/backend/emitter/SJSGen.scala
@@ -16,6 +16,7 @@ import org.scalajs.ir._
 import org.scalajs.ir.Names._
 import org.scalajs.ir.Types._
 import org.scalajs.ir.{Trees => irt}
+import org.scalajs.ir.WellKnownNames._
 
 import org.scalajs.linker.backend.javascript.Trees._
 import org.scalajs.linker.interface._
