+++ linker/shared/src/main/scala/org/scalajs/linker/backend/wasmemitter/Preprocessor.scala
@@ -17,6 +17,7 @@ import scala.collection.mutable
 import org.scalajs.ir.Names._
 import org.scalajs.ir.Trees._
 import org.scalajs.ir.Types._
+import org.scalajs.ir.WellKnownNames._
 import org.scalajs.ir.{ClassKind, Traversers}
 
 import org.scalajs.linker.standard.{CoreSpec, LinkedClass, LinkedTopLevelExport}
