+++ linker/shared/src/main/scala/org/scalajs/linker/checker/IRChecker.scala
@@ -20,6 +20,7 @@ import org.scalajs.ir._
 import org.scalajs.ir.Names._
 import org.scalajs.ir.Trees._
 import org.scalajs.ir.Types._
+import org.scalajs.ir.WellKnownNames._
 
 import org.scalajs.logging._
