+++ linker/shared/src/main/scala/org/scalajs/linker/frontend/MethodSynthesizer.scala
@@ -22,6 +22,7 @@ import org.scalajs.ir
 import org.scalajs.ir.Names._
 import org.scalajs.ir.Trees._
 import org.scalajs.ir.Types._
+import org.scalajs.ir.WellKnownNames._
 
 import Analysis._
