+++ linker/shared/src/test/scala/org/scalajs/linker/BaseLinkerTest.scala
@@ -19,6 +19,7 @@ import org.scalajs.ir.ClassKind
 import org.scalajs.ir.Names._
 import org.scalajs.ir.Trees._
 import org.scalajs.ir.Types._
+import org.scalajs.ir.WellKnownNames._
 
 import org.scalajs.junit.async._
