+++ linker/shared/src/test/scala/org/scalajs/linker/FewestModulesSplittingTest.scala
@@ -18,9 +18,9 @@ import org.junit.Test
 import org.junit.Assert._
 
 import org.scalajs.ir.ClassKind
-import org.scalajs.ir.Names._
 import org.scalajs.ir.Trees._
 import org.scalajs.ir.Types._
+import org.scalajs.ir.WellKnownNames._
 
 import org.scalajs.junit.async._
