+++ linker/shared/src/test/scala/org/scalajs/linker/IRCheckerTest.scala
@@ -23,6 +23,7 @@ import org.scalajs.ir.Names._
 import org.scalajs.ir.Transformers._
 import org.scalajs.ir.Trees._
 import org.scalajs.ir.Types._
+import org.scalajs.ir.WellKnownNames._
 
 import org.scalajs.logging._
