+++ linker/shared/src/test/scala/org/scalajs/linker/IncrementalTest.scala
@@ -24,6 +24,7 @@ import org.scalajs.ir.Names._
 import org.scalajs.ir.Trees._
 import org.scalajs.ir.Types._
 import org.scalajs.ir.Version
+import org.scalajs.ir.WellKnownNames._
 
 import org.scalajs.logging._
