+++ linker/shared/src/test/scala/org/scalajs/linker/testutils/TestIRBuilder.scala
@@ -22,6 +22,7 @@ import org.scalajs.ir.OriginalName
 import org.scalajs.ir.OriginalName.NoOriginalName
 import org.scalajs.ir.Trees._
 import org.scalajs.ir.Types._
+import org.scalajs.ir.WellKnownNames._
 
 import org.scalajs.linker.interface.ModuleInitializer
