+++ project/JavalibIRCleaner.scala
@@ -6,6 +6,7 @@ import org.scalajs.ir.Names._
 import org.scalajs.ir.Trees._
 import org.scalajs.ir.Types._
 import org.scalajs.ir.Version.Unversioned
+import org.scalajs.ir.WellKnownNames._
 
 import java.io._
 import java.net.URI
